Added Archer class Added Archer First Skill (Focus) - chance to double attack damage per standard attack (PASSIVE) Added Archer Second Skill (StraightShot) - Fires an explosive arrow Fixed Trainer bug (Mobs would attack it and crash the server) Changed Weapon Shapes to switch between lib files. 0-99 Standard weapons. 100-199 Assassin Weapons. 200-299 Archer Weapons.

+++ trunk/Server/MirObjects/Monsters/HumanWizard.cs
@@ -161,8 +161,8 @@ namespace Server.MirObjects.Monsters
         public override Packet GetInfo()
         {
             PlayerObject master = null;
-            sbyte weapon = -1;
-            sbyte armour = 0;
+            short weapon = -1;
+            short armour = 0;
             byte wing = 0;
             if (Master != null && Master is PlayerObject) master = (PlayerObject)Master;
             if (master != null)
